+++ desktop/electron/preload.ts
@@ -1,6 +1,5 @@
-import { contextBridge, ipcRenderer } from 'electron'
+import { contextBridge } from 'electron'
 
 contextBridge.exposeInMainWorld('workbench', {
   platform: process.platform,
-  getSystemLocale: (): Promise<string> => ipcRenderer.invoke('workbench:get-system-locale'),
 })